ISSUE #768 improved test coverage and ntasks logic
